[BetterZip] Update Raycast Api (#13374)
* [BetterZip] Update Raycast Api

* Update CHANGELOG.md and optimise images

---------
