+++ extensions/betterzip/README.md
@@ -3,8 +3,3 @@
 Compress and extract selected items in Finder with BetterZip via Raycast.
 
 To use this command you need to have BetterZip installed on your Mac. [Download BetterZip](https://macitbetter.com).
-
-**TODO:** More commands
-
-- Compress Files With Presets
-- Extract Files With Presets
