+++ extensions/betterzip/src/open-with-betterzip.tsx
@@ -1,6 +1,5 @@
-import { closeMainWindow, getSelectedFinderItems, open, showHUD } from "@raycast/api";
-import { betterZipInstalled } from "./utils/common-utils";
-import { betterZipNotInstallDialog } from "./hooks/hooks";
+import { closeMainWindow, getSelectedFinderItems, open, showToast, Toast } from "@raycast/api";
+import { betterZipInstalled, betterZipNotInstallDialog } from "./utils/common-utils";
 
 export default async () => {
   try {
@@ -11,7 +10,7 @@ export default async () => {
     await closeMainWindow({ clearRootSearch: false });
     const fileSystemItems = await getSelectedFinderItems();
     if (fileSystemItems.length === 0) {
-      await showHUD("No files selected");
+      await showToast({ title: "No files selected", style: Toast.Style.Failure });
       return;
     }
     fileSystemItems.forEach((value) => {
@@ -19,6 +18,6 @@ export default async () => {
     });
   } catch (e) {
     console.error(String(e));
-    await showHUD(String(e));
+    await showToast({ title: String(e), style: Toast.Style.Failure });
   }
 };
