+++ extensions/betterzip/src/utils/common-utils.ts
@@ -1,4 +1,5 @@
 import * as fs from "fs";
+import { Alert, confirmAlert, open } from "@raycast/api";
 
 export const betterZipInstalled = () => {
   try {
@@ -9,4 +10,19 @@ export const betterZipInstalled = () => {
     console.error(String(e));
     return false;
   }
+};
+
+export const betterZipNotInstallDialog = async () => {
+  const options: Alert.Options = {
+    icon: { source: "betterzip-icon.png" },
+    title: "BetterZip Not Installed",
+    message: "BetterZip is not installed on your Mac. Please install BetterZip to use this command.",
+    primaryAction: {
+      title: "Get BetterZip",
+      onAction: () => {
+        open("https://macitbetter.com");
+      },
+    },
+  };
+  await confirmAlert(options);
 };
